chore: Refactoring ColumnSizing into folder and separate base, types, and utils
