+++ packages/table-core/src/core/table.ts
@@ -28,7 +28,7 @@ import { ColumnFiltering } from '../features/column-filtering/ColumnFiltering'
 import { ColumnGrouping } from '../features/column-grouping/ColumnGrouping'
 import { ColumnOrdering } from '../features/column-ordering/ColumnOrdering'
 import { ColumnPinning } from '../features/column-pinning/ColumnPinning'
-import { ColumnSizing } from '../features/ColumnSizing'
+import { ColumnSizing } from '../features/column-sizing/ColumnSizing'
 import { ColumnVisibility } from '../features/column-visibility/ColumnVisibility'
 import { GlobalFaceting } from '../features/GlobalFaceting'
 import { GlobalFiltering } from '../features/GlobalFiltering'
