+++ packages/table-core/src/features/column-sizing/isTouchStartEvent.ts
@@ -1,0 +1,3 @@
+export function isTouchStartEvent(e: unknown): e is TouchEvent {
+  return (e as TouchEvent).type === 'touchstart'
+}
