+++ packages/table-core/src/features/column-sizing/passiveSupported.ts
@@ -1,0 +1,24 @@
+export function passiveEventSupported() {
+  let passiveSupported: boolean | null = null
+
+  if (typeof passiveSupported === 'boolean') return passiveSupported
+
+  let supported = false
+  try {
+    const options = {
+      get passive() {
+        supported = true
+        return false
+      },
+    }
+
+    const noop = () => {}
+
+    window.addEventListener('test', noop, options)
+    window.removeEventListener('test', noop)
+  } catch (err) {
+    supported = false
+  }
+  passiveSupported = supported
+  return passiveSupported
+}
